+++ packages/contracts/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add a `SubtreeUpdate` event to `CommitmentTreeManager` for when a subtree update is committed 
 - Add `test:gas-report` command
+- Add batch proof verification
 - Migrate for joinsplit
   - Compute operationDigest once for the entire operation
 - Rename all "flax" instances to "nocturne"